+++ opyoid/bindings/multi_binding/list_provider.py
@@ -1,7 +1,7 @@
 from typing import List
 
 from opyoid.provider import Provider
-from opyoid.typings import InjectedT
+from opyoid.utils import InjectedT
 
 
 class ListProvider(Provider[List[InjectedT]]):