+++ opyoid/bindings/self_binding/self_binding.py
@@ -4,10 +4,10 @@ import attr
 
 from opyoid.bindings.binding import Binding
 from opyoid.scopes import Scope, SingletonScope
-from opyoid.typings import InjectedT
+from opyoid.utils import InjectedT
 
 
-@attr.s(auto_attribs=True, frozen=True)
+@attr.s(auto_attribs=True, frozen=True, repr=False)
 class SelfBinding(Binding[InjectedT]):
     _target_type: Type[InjectedT]
     scope: Type[Scope] = SingletonScope
@@ -20,3 +20,7 @@ class SelfBinding(Binding[InjectedT]):
     @property
     def annotation(self) -> Optional[str]:
         return self._annotation
+
+    def __repr__(self) -> str:
+        scope_string = f", scope={self.scope}" if self.scope != SingletonScope else ""
+        return f"{self.__class__.__name__}({self.target!r}{scope_string})"