+++ opyoid/providers/provider_creator.py
@@ -5,7 +5,7 @@ from typing import List
 from opyoid.exceptions import NoBindingFound
 from opyoid.injection_context import InjectionContext
 from opyoid.provider import Provider
-from opyoid.typings import InjectedT
+from opyoid.utils import InjectedT
 from .providers_factories import FromBindingProviderFactory, FromCacheProviderFactory, OptionalProviderFactory, \
     ProviderFactory, ProviderProviderFactory, SetProviderFactory, TupleProviderFactory, TypeProviderFactory
 from .providers_factories.jit_provider_factory import JitProviderFactory
@@ -41,4 +41,4 @@ class ProviderCreator:
         for provider_factory in self._provider_factories:
             if provider_factory.accept(context):
                 return provider_factory.create(context)
-        raise NoBindingFound(f"Could not find any bindings for {context.target}")
+        raise NoBindingFound(f"Could not find any bindings for {context.target!r}")