+++ opyoid/providers/providers_factories/from_cache_provider_factory.py
@@ -1,6 +1,6 @@
 from opyoid.injection_context import InjectionContext
 from opyoid.provider import Provider
-from opyoid.typings import InjectedT
+from opyoid.utils import InjectedT
 from .provider_factory import ProviderFactory