+++ opyoid/providers/providers_factories/provider_provider_factory.py
@@ -3,7 +3,7 @@ from opyoid.injection_context import InjectionContext
 from opyoid.provider import Provider
 from opyoid.target import Target
 from opyoid.type_checker import TypeChecker
-from opyoid.typings import InjectedT
+from opyoid.utils import InjectedT
 from .provider_factory import ProviderFactory