+++ opyoid/typings.py
@@ -1,4 +1,0 @@
-from typing import Any, TypeVar
-
-InjectedT = TypeVar("InjectedT", bound=Any)
-EMPTY = object()